requirements.txt
ollama
fastapi
uvicorn
streamlit
streamlit-lottie
requests
pyautogui
mss
Pillow
pyinstaller
SpeechRecognition
pyttsx3
chromadb
playwright
peft
psutil
supabase
python-dotenv
duckduckgo-search
crewai
openai
pystray
pywebview
pynput
docker
stripe
pyperclip
watchdog
rich
cryptography
